# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Test Task",
        "contact": {
            "email": "mahmoudnada5050@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "mahmoud"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "mahmoudnada5050@gmail.com"
                    },
                    {
                        "name": "device_token",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "device_token"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "mahmoudnada5050@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "logout currently logged in user",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "refreshes expired token",
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/me": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Get the current logged in user",
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/update": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My Account",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "mahmoud"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "mahmoudnada5050@gmail.com"
                    },
                    {
                        "name": "photo",
                        "in": "formData",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/update-password": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My Password",
                "parameters": [
                    {
                        "name": "current_password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Model"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/update-device-token": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My device token",
                "parameters": [
                    {
                        "name": "device_token",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Model"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/notifications": {
            "get": {
                "tags": [
                    "notifications"
                ],
                "summary": "Get all tips notifications paginate 5 per page",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Langouge Code",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "en"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/notifications/{id}": {
            "get": {
                "tags": [
                    "notifications"
                ],
                "summary": "read notification",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Langouge Code",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "en"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/read-all-notifications": {
            "get": {
                "tags": [
                    "notifications"
                ],
                "summary": "read-all-notifications",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Langouge Code",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "en"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}